requirements.txt
streamlit
pandas
numpy
yfinance
plotly
scikit-learn
sentence-transformers
umap-learn
matplotlib